requirements.txt
Django==1.9.6
django-braces==1.9.0
django-filter==0.13.0
django-guardian==1.4.4
django-widget-tweaks==1.4.1
django-youtube==0.2
gdata==2.0.18
google-api-python-client==1.5.0
httplib2==0.9.2
oauth2client==2.0.2
Pillow==3.2.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
